sentinel_lambda/requirements.txt
aws-cdk-lib==2.15.0
constructs>=10.0.0,<11.0.0
cdk-nag==2.4.10
dacite==1.6.0
packaging==21.3
